update driver company style
