+++ src/components/DriversCompaniesList/styles.js
@@ -103,14 +103,6 @@ export const CheckBoxWrapper = styled.div`
   `}
 `
 
-export const ActionsContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-top: 8px;
-  padding-bottom: 8px;
-`
-
 export const EnableWrapper = styled.div`
   display: flex;
   align-items: center;
@@ -125,56 +117,6 @@ export const EnableWrapper = styled.div`
     ` : css`
       padding-right: 10px;
     `}
-  }
-`
-
-export const ActionSelectorWrapper = styled.div`
-  ${props => props.theme?.rtl ? css`
-    margin-right: 20px;
-  ` : css`
-    margin-left: 20px;
-  `}
-  button {
-    background: transparent !important;
-    border: none;
-    padding: 0px;
-    &:active,
-    &:focus {
-      border-color: unset !important;
-      box-shadow: none !important;
-    }
-    svg {
-      color: ${props => props.theme.colors.headingColor};
-    }
-
-    &:after {
-      display: none;
-    }
-  }
-
-  .show {
-    button {
-      background: ${props => props.theme.colors.secundary} !important;
-      border-color: unset !important;
-      box-shadow: none !important;
-    }
-    >div {
-      border: 1px solid ${props => props.theme.colors.borderColor};
-      box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.12);
-    }
-  }
-
-  > div {
-    > div {
-      border-radius: 8px;
-      .dropdown-item {
-        font-size: 12px;
-        color: ${props => props.theme.colors.headingColor};
-      }
-      .dropdown-item:last-child {
-        color: #E63757;
-      }
-    }
   }
 `
