+++ src/components/DriversCompanyDetailsForm/index.js
@@ -3,17 +3,21 @@ import {
   useLanguage,
   DriversCompanyDetails as DriversCompanyDetailsController
 } from 'ordering-components-admin'
-import { Alert } from '../Confirm'
+import { Alert, Confirm } from '../Confirm'
 import { DragScroll } from '../DragScroll'
 import { DriversCompanyGeneralDetails } from '../DriversCompanyGeneralDetails'
 import { DriversCompanyScheduleDetails } from '../DriversCompanyScheduleDetails'
 import { DriversCompanyWebhooksDetails } from '../DriversCompanyWebhooksDetails'
+import { Dropdown, DropdownButton } from 'react-bootstrap'
+import { ThreeDots } from 'react-bootstrap-icons'
+import { useTheme } from 'styled-components'
 
 import {
   DetailsContainer,
   Header,
   TabsContainer,
-  Tab
+  Tab,
+  ActionSelectorWrapper
 } from './styles'
 
 const DriversCompanyDetailsFormUI = (props) => {
@@ -21,12 +25,15 @@ const DriversCompanyDetailsFormUI = (props) => {
     driversCompany,
     changesState,
     actionState,
-    cleanActionState
+    cleanActionState,
+    handleDeleteDriversCompany
   } = props
 
   const [, t] = useLanguage()
+  const theme = useTheme()
   const [currentTabItem, setCurrentTabItem] = useState('general')
   const [alertState, setAlertState] = useState({ open: false, content: [] })
+  const [confirm, setConfirm] = useState({ open: false, content: null, handleOnAccept: null })
 
   const tabItems = [
     { key: 'general', content: t('GENERAL', 'General') },
@@ -39,6 +46,17 @@ const DriversCompanyDetailsFormUI = (props) => {
     setAlertState({
       open: false,
       content: []
+    })
+  }
+
+  const onDeleteCompany = () => {
+    setConfirm({
+      open: true,
+      content: t('QUESTION_DELETE_DRIVER_COMPANY', 'Are you sure to remove this driver company?'),
+      handleOnAccept: () => {
+        setConfirm({ ...confirm, open: false })
+        handleDeleteDriversCompany()
+      }
     })
   }
 
@@ -62,6 +80,21 @@ const DriversCompanyDetailsFormUI = (props) => {
                 : t('DRIVER_COMPANY_SETTINGS', 'Driver company settings')
             }
           </h1>
+          {driversCompany && (
+            <ActionSelectorWrapper>
+              <DropdownButton
+                menuAlign={theme?.rtl ? 'left' : 'right'}
+                title={<ThreeDots />}
+                id={theme?.rtl ? 'dropdown-menu-align-left' : 'dropdown-menu-align-right'}
+              >
+                <Dropdown.Item
+                  onClick={() => onDeleteCompany()}
+                >
+                  {t('DELETE', 'Delete')}
+                </Dropdown.Item>
+              </DropdownButton>
+            </ActionSelectorWrapper>
+          )}
         </Header>
         <TabsContainer>
           <DragScroll>
@@ -95,6 +128,16 @@ const DriversCompanyDetailsFormUI = (props) => {
         open={alertState.open}
         onClose={() => closeAlert()}
         onAccept={() => closeAlert()}
+        closeOnBackdrop={false}
+      />
+      <Confirm
+        title={t('WEB_APPNAME', 'Ordering')}
+        content={confirm.content}
+        acceptText={t('ACCEPT', 'Accept')}
+        open={confirm.open}
+        onClose={() => setConfirm({ ...confirm, open: false })}
+        onCancel={() => setConfirm({ ...confirm, open: false })}
+        onAccept={confirm.handleOnAccept}
         closeOnBackdrop={false}
       />
     </>
